# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ContextHub API
  version: 0.9.0
  description: Agent-native context backend API.
servers:
  - url: http://127.0.0.1:4040
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
    RecordLayer:
      type: string
      enum: [l0, l1, l2]

    DeriveMode:
      type: string
      enum: [sync, async]

    DeriveStrategy:
      type: string
      enum: [preserve_manual, create_sidecar, replace_derived_only]

    PromptPreset:
      type: string
      enum: [archive_and_memory, memory_only, archive_only]

    ImportContentKind:
      type: string
      enum: [inline_text, markdown_file, external_ref, blob_ref]

    Tenant:
      type: object
      additionalProperties: true
      required: [id, slug, name, createdAt]
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time

    CreateTenantRequest:
      type: object
      required: [slug, name]
      properties:
        slug:
          type: string
        name:
          type: string
        description:
          type: string

    CreatePartitionRequest:
      type: object
      required: [tenantId, key, name]
      properties:
        tenantId:
          type: string
        key:
          type: string
        name:
          type: string
        kind:
          type: string
        description:
          type: string
        allowCrossQueryFrom:
          type: array
          items:
            type: string

    RegisterAgentRequest:
      type: object
      required: [tenantId, name]
      properties:
        tenantId:
          type: string
        name:
          type: string
        kind:
          type: string
        metadata:
          type: object
          additionalProperties: true

    CreatePrincipalRequest:
      type: object
      required: [tenantId, name]
      properties:
        tenantId:
          type: string
        name:
          type: string
        kind:
          type: string
        metadata:
          type: object
          additionalProperties: true

    UpsertPrincipalAclRequest:
      type: object
      required: [partitionKey]
      properties:
        partitionKey:
          type: string
        canRead:
          type: boolean
        canWrite:
          type: boolean
        allowedLayers:
          type: array
          items:
            $ref: "#/components/schemas/RecordLayer"

    Record:
      type: object
      additionalProperties: true
      required: [id, tenantId, partitionKey, type, layer, title, text, createdAt, updatedAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        partitionKey:
          type: string
        type:
          type: string
        layer:
          $ref: "#/components/schemas/RecordLayer"
        title:
          type: string
        text:
          type: string
        source:
          oneOf:
            - type: object
              additionalProperties: true
            - type: "null"
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
        manualSummary:
          type: string
        importance:
          type: number
        pinned:
          type: boolean
        idempotencyKey:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateRecordRequest:
      type: object
      required: [tenantId, partitionKey, type, title, text]
      properties:
        tenantId:
          type: string
        partitionKey:
          type: string
        type:
          type: string
        layer:
          $ref: "#/components/schemas/RecordLayer"
        title:
          type: string
        text:
          type: string
        manualSummary:
          type: string
        source:
          oneOf:
            - type: object
              additionalProperties: true
            - type: "null"
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
        importance:
          type: number
          minimum: 0
          maximum: 5
        pinned:
          type: boolean
        idempotencyKey:
          type: string

    UpdateRecordRequest:
      type: object
      properties:
        type:
          type: string
        layer:
          $ref: "#/components/schemas/RecordLayer"
        title:
          type: string
        text:
          type: string
        manualSummary:
          type: string
        source:
          oneOf:
            - type: object
              additionalProperties: true
            - type: "null"
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
        importance:
          type: number
          minimum: 0
          maximum: 5
        pinned:
          type: boolean

    GrepRequest:
      type: object
      required: [tenantId, pattern]
      properties:
        tenantId:
          type: string
        pattern:
          type: string
        partitions:
          type: array
          items:
            type: string
        types:
          type: array
          items:
            type: string
        layers:
          type: array
          items:
            $ref: "#/components/schemas/RecordLayer"
        regex:
          type: boolean
        caseSensitive:
          type: boolean
        limit:
          type: integer
        beforeContext:
          type: integer
        afterContext:
          type: integer

    ImportContent:
      type: object
      required: [kind]
      properties:
        kind:
          $ref: "#/components/schemas/ImportContentKind"
        text:
          type: string
        path:
          type: string
        url:
          type: string

    DeriveOptions:
      type: object
      properties:
        enabled:
          type: boolean
        mode:
          $ref: "#/components/schemas/DeriveMode"
        emitLayers:
          type: array
          items:
            $ref: "#/components/schemas/RecordLayer"
        strategy:
          $ref: "#/components/schemas/DeriveStrategy"
        promptPreset:
          $ref: "#/components/schemas/PromptPreset"
        provider:
          type: string
        model:
          type: string

    ImportResourceRequest:
      type: object
      required: [tenantId, partitionKey, targetLayer, title, content]
      properties:
        tenantId:
          type: string
        partitionKey:
          type: string
        type:
          type: string
        targetLayer:
          $ref: "#/components/schemas/RecordLayer"
        title:
          type: string
        content:
          $ref: "#/components/schemas/ImportContent"
        source:
          oneOf:
            - type: object
              additionalProperties: true
            - type: "null"
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
        manualSummary:
          type: string
        importance:
          type: number
        pinned:
          type: boolean
        idempotencyKey:
          type: string
        derive:
          $ref: "#/components/schemas/DeriveOptions"

    QueryRequest:
      type: object
      required: [tenantId, query]
      properties:
        tenantId:
          type: string
        query:
          type: string
        partitions:
          type: array
          items:
            type: string
        types:
          type: array
          items:
            type: string
        layers:
          type: array
          items:
            $ref: "#/components/schemas/RecordLayer"
        limit:
          type: integer
        rerank:
          type: boolean

    QueryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
        retrieval:
          type: object
          additionalProperties: true

    DerivationJob:
      type: object
      additionalProperties: true
      required: [id, tenantId, partitionKey, sourceRecordId, status, mode, requestedLayers, provider, promptPreset, createdAt, updatedAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        partitionKey:
          type: string
        sourceRecordId:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed]
        mode:
          $ref: "#/components/schemas/DeriveMode"
        effectiveMode:
          oneOf:
            - $ref: "#/components/schemas/DeriveMode"
            - type: "null"
        requestedLayers:
          type: array
          items:
            $ref: "#/components/schemas/RecordLayer"
        provider:
          type: string
        model:
          type: string
        promptPreset:
          $ref: "#/components/schemas/PromptPreset"
        errorMessage:
          oneOf:
            - type: string
            - type: "null"
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        finishedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"

    RecordLink:
      type: object
      additionalProperties: true
      required: [id, tenantId, sourceRecordId, targetRecordId, relation, metadata, createdAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        sourceRecordId:
          type: string
        targetRecordId:
          type: string
        relation:
          type: string
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time

    CommitSessionRequest:
      type: object
      required: [tenantId, partitionKey]
      properties:
        tenantId:
          type: string
        partitionKey:
          type: string
        agentId:
          type: string
        sessionId:
          type: string
        summary:
          type: string
        metadata:
          type: object
          additionalProperties: true
        messages:
          type: array
          items:
            type: object
            additionalProperties: true
        memoryEntries:
          type: array
          items:
            type: object
            additionalProperties: true

paths:
  /health:
    get:
      summary: Service health and provider readiness
      responses:
        "200":
          description: OK

  /v1/auth/me:
    get:
      security:
        - bearerAuth: []
      summary: Return current auth identity
      responses:
        "200":
          description: OK

  /v1/tenants:
    post:
      security:
        - bearerAuth: []
      summary: Create or return a tenant by slug
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"

  /v1/partitions:
    post:
      security:
        - bearerAuth: []
      summary: Create or return a partition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePartitionRequest"
      responses:
        "201":
          description: Created

  /v1/agents:
    post:
      security:
        - bearerAuth: []
      summary: Register an agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterAgentRequest"
      responses:
        "201":
          description: Created

  /v1/principals:
    post:
      security:
        - bearerAuth: []
      summary: Create a principal and return its token once
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrincipalRequest"
      responses:
        "201":
          description: Created

  /v1/principals/{principalId}/acl:
    post:
      security:
        - bearerAuth: []
      summary: Create or update partition ACL for a principal
      parameters:
        - name: principalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertPrincipalAclRequest"
      responses:
        "201":
          description: Created

  /v1/records:
    post:
      security:
        - bearerAuth: []
      summary: Write a curated record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecordRequest"
      responses:
        "201":
          description: Created

  /v1/records/{recordId}:
    get:
      security:
        - bearerAuth: []
      summary: Fetch one record
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Record"
    patch:
      security:
        - bearerAuth: []
      summary: Update one record in place
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecordRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Record"

  /v1/records/{recordId}/lines:
    get:
      security:
        - bearerAuth: []
      summary: Read one record as numbered lines
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: string
        - name: from_line
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 80
      responses:
        "200":
          description: OK

  /v1/records/grep:
    post:
      security:
        - bearerAuth: []
      summary: Search record text line-by-line with line numbers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrepRequest"
      responses:
        "200":
          description: OK

  /v1/resources/import:
    post:
      security:
        - bearerAuth: []
      summary: Import resource, optionally derive layers sync/async, and persist lineage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportResourceRequest"
      responses:
        "201":
          description: Created

  /v1/derivation-jobs/{jobId}:
    get:
      security:
        - bearerAuth: []
      summary: Fetch a derivation job and its current status
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DerivationJob"

  /v1/records/{recordId}/links:
    get:
      security:
        - bearerAuth: []
      summary: List record links for a source record
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/RecordLink"

  /v1/query:
    post:
      security:
        - bearerAuth: []
      summary: Query records with optional rerank
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"

  /v1/sessions/commit:
    post:
      security:
        - bearerAuth: []
      summary: Commit a session and optional memory entries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitSessionRequest"
      responses:
        "201":
          description: Created